requirements.txt
alembic==1.15.2
blinker==1.9.0
certifi==2025.4.26
charset-normalizer==3.4.2
click==8.2.0
Flask==3.1.1
Flask-Login==0.6.3
Flask-Migrate==4.1.0
Flask-SQLAlchemy==3.1.1
greenlet==3.2.2
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.6
Mako==1.3.10
MarkupSafe==3.0.2
psycopg2-binary==2.9.10
python-dotenv==1.1.0
pytz==2025.2
requests==2.32.3
SQLAlchemy==2.0.41
typing_extensions==4.13.2
urllib3==2.4.0
Werkzeug==3.1.3
